Moved all files in teh server folder into the serber/user folder to prepare to add a url shortner
