+++ src/routes/[name]/+page.server.ts
@@ -1,4 +1,4 @@
-import { User_Model } from '$lib/server/models';
+import { User_Model } from '$lib/server/user/models';
 import type { Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
