+++ src/routes/[name]/games/[id]/+page.server.ts
@@ -1,4 +1,4 @@
-import { User_Model } from '$lib/server/models';
+import { User_Model } from '$lib/server/user/models';
 
 import type { Types } from 'mongoose';
 import type { PageServerLoad } from '../$types';
