+++ src/routes/dashboard/+page.server.ts
@@ -1,4 +1,4 @@
-import { User_Model } from '$lib/server/models';
+import { User_Model } from '$lib/server/user/models';
 
 export async function load(event): Promise<unknown> {
 	const email = event.cookies.get('email');
