+++ src/routes/dashboard/games/Precision/One-Player/+page.server.ts
@@ -1,5 +1,5 @@
 // Import necessary modules and types
-import { User_Model } from '$lib/server/models';
+import { User_Model } from '$lib/server/user/models';
 import { getTimeThreshold } from '$lib/shared/utils';
 import type { Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from '../$types';
